Reformat start_exec.c - I got rid of about 10k tabs and replaced them with spaces, as well as doing some other reformatting things to get the style under control in one of our ugliest files

+++ src/include/server.h
@@ -277,7 +277,7 @@ extern int svr_save(struct server *, int);
 #endif /* END MAX */
 
 #ifndef JobStatRate
-#define JobStatRate MAX(PBS_JOBSTAT_MIN,server.sv_attr[(int)SRV_ATR_JobStatRate].at_val.at_long)
+#define JobStatRate MAX(PBS_JOBSTAT_MIN,server.sv_attr[SRV_ATR_JobStatRate].at_val.at_long)
 #endif
